Move DCR Config util methods to Service layer
