+++ components/org.wso2.carbon.identity.oauth.dcr/src/test/resources/testng.xml
@@ -31,7 +31,6 @@
             <class name="org.wso2.carbon.identity.oauth.dcr.factory.RegistrationRequestFactoryTest"/>
             <class name="org.wso2.carbon.identity.oauth.dcr.factory.UnregistrationRequestFactoryTest"/>
             <class name="org.wso2.carbon.identity.oauth.dcr.util.DCRMUtilsTest"/>
-            <class name="org.wso2.carbon.identity.oauth.dcr.util.DCRConfigUtilsTest"/>
             <class name="org.wso2.carbon.identity.oauth.dcr.handler.RegistrationHandlerTest"/>
             <class name="org.wso2.carbon.identity.oauth.dcr.handler.UnRegistrationHandlerTest"/>
             <class name="org.wso2.carbon.identity.oauth.dcr.factory.HttpUnRegistrationResponseFactoryTest"/>
@@ -52,7 +51,6 @@
         <class name="org.wso2.carbon.identity.oauth.dcr.factory.HttpUnRegistrationResponseFactoryTest"/>
         <class name="org.wso2.carbon.identity.oauth.dcr.util.DCRMUtilsTest"/>
         <class name="org.wso2.carbon.identity.oauth.dcr.util.HandlerManagerTest"/>
-        <class name="org.wso2.carbon.identity.oauth.dcr.util.DCRConfigUtilsTest"/>
         <class name="org.wso2.carbon.identity.oauth.dcr.handler.RegistrationHandlerTest"/>
         <class name="org.wso2.carbon.identity.oauth.dcr.handler.UnRegistrationHandlerTest"/>
         <class name="org.wso2.carbon.identity.oauth.dcr.DCRConfigurationMgtServiceImplTest"/>
